# Conflicts:
#	pubspec.yaml

+++ lib/screens/client_screen/services/model/lawyer_list_model.dart
@@ -50,7 +50,7 @@ class Data {
   List<String>? languageSpoken;
   String? experience;
   String? image;
-  int? avgRating;
+  double? avgRating;
 
   // double? avgRating;
